Refactored ship logic into separate module
